# Conflicts:
#	.idea/caches/build_file_checksums.ser
#	app/src/main/java/kr/or/dgit/it/cosmeticmngapp/MyItemList.java

전체 선택 전체 해제

+++ app/src/main/java/kr/or/dgit/it/cosmeticmngapp/MainActivity.java
@@ -3,55 +3,35 @@ package kr.or.dgit.it.cosmeticmngapp;
 import android.Manifest;
 import android.app.Activity;
 import android.app.ActivityManager;
-import android.app.AlarmManager;
 import android.app.Fragment;
-import android.app.NotificationManager;
-import android.app.PendingIntent;
 import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.pm.PackageManager;
-import android.database.Cursor;
-import android.graphics.Bitmap;
-import android.graphics.BitmapFactory;
 import android.os.Bundle;
-import android.os.Handler;
-import android.os.Message;
 import android.os.Handler;
 import android.os.Message;
 import android.support.design.widget.BottomNavigationView;
 import android.support.design.widget.NavigationView;
 import android.support.v4.app.ActivityCompat;
-
-import android.support.v4.app.NotificationCompat;
 import android.support.v4.content.ContextCompat;
 import android.support.v4.widget.DrawerLayout;
 import android.support.v7.app.ActionBarDrawerToggle;
 import android.support.v7.app.AlertDialog;
+import android.support.v7.view.menu.ActionMenuItemView;
 import android.support.v7.widget.Toolbar;
 import android.util.Log;
-import android.view.Gravity;
-import android.view.LayoutInflater;
 import android.view.Menu;
-import android.view.MenuInflater;
-import android.view.MenuItem;
 import android.view.View;
 import android.widget.Button;
-import android.widget.CheckBox;
-import android.widget.CompoundButton;
 import android.widget.TextView;
 import android.widget.Toast;
 
-import java.text.SimpleDateFormat;
 import java.util.ArrayList;
-import java.util.Calendar;
-import java.util.Date;
 
 import kr.or.dgit.it.cosmeticmngapp.dao.UserCosmeticDAO;
 import kr.or.dgit.it.cosmeticmngapp.dao.UserCosmeticToolsDAO;
 import kr.or.dgit.it.cosmeticmngapp.dao.UserLensDAO;
 import kr.or.dgit.it.cosmeticmngapp.db.DBhelper;
-
-import static kr.or.dgit.it.cosmeticmngapp.R.string.addAlldel;
 
 public class MainActivity extends android.support.v7.app.AppCompatActivity implements NavigationView.OnNavigationItemSelectedListener {
     private static final String TAG = "Main";
@@ -69,13 +49,29 @@ public class MainActivity extends android.support.v7.app.AppCompatActivity imple
     UserCosmeticToolsDAO userCosmeticToolsDAO;
     ArrayList<Integer> numberlist = new ArrayList<>();
     int favorite = 0;
+    private Toolbar toolbar;
+    Handler checkedhandler;
+
+    public static MainActivity newInstance() {
+        return  new MainActivity();
+    }
+
+    public Handler getCheckedhandler() {
+        return checkedhandler;
+    }
+
+    public void setCheckedhandler(Handler checkedhandler) {
+        this.checkedhandler = checkedhandler;
+    }
+
+
 
     @Override
     protected void onCreate(android.os.Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         Log.d(TAG, "onCreate: ");
         setContentView(R.layout.activity_main);
-        Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
+        toolbar = (Toolbar) findViewById(R.id.toolbar);
         setSupportActionBar(toolbar);
 
         userCosmeticDAO = new UserCosmeticDAO(this);
@@ -86,11 +82,6 @@ public class MainActivity extends android.support.v7.app.AppCompatActivity imple
         actionBar.setDisplayShowCustomEnabled(true);
         actionBar.setDisplayShowTitleEnabled(true);
         actionBar.setDisplayHomeAsUpEnabled(true);
-        // getSupportActionBar().setHomeAsUpIndicator(R.drawable.menu);
-
-        //  toolbar.setNavigationIcon(R.drawable.menu);
-
-        setTitleName();
 
         drawerLayout = findViewById(R.id.drawerlayout);
         ActionBarDrawerToggle toggle = new ActionBarDrawerToggle(this, drawerLayout, toolbar, R.string.openT, R.string.closeT);
@@ -100,15 +91,19 @@ public class MainActivity extends android.support.v7.app.AppCompatActivity imple
         android.support.design.widget.NavigationView navigationView = findViewById(R.id.navigation);
         navigationView.setNavigationItemSelectedListener(this);
 
-        fragment = MyItemList.newInstance();
+        //fragNum = 1;
+        MyItemList fragment = MyItemList.newInstance();
         fragment.setToolbarHandler(toolbarHandler);
         Bundle bundle = new Bundle();
-        bundle.putInt("frag", 1);
+        bundle.putInt("frag", fragNum);
         fragment.setArguments(bundle);
-        fragment.getListDatas();
+        //fragment.getListDatas();
+
+        setTitleName();
 
         emptyTV = (TextView) findViewById(R.id.emptyTV);
         getFragmentManager().beginTransaction().replace(R.id.fragment_container, fragment).commit();
+
         Button closeNavi = navigationView.findViewById(R.id.naviCloseBtn);
         /* closeNavi.setOnClickListener();*/
 
@@ -175,27 +170,28 @@ public class MainActivity extends android.support.v7.app.AppCompatActivity imple
         setTitleName();
 
         fragment = MyItemList.newInstance();
-        fragment.setToolbarHandler(toolbarHandler);
-        Bundle bundle = new Bundle();
-        bundle.putInt("frag", fragNum);
-        fragment.setArguments(bundle);
-        Log.d("nasdfasdfsdfnnn","졸라망함............1");
-        fragment.getListDatas();
-
-
+        ActionMenuItemView menuItem1 = findViewById(R.id.searchIcon);
+        /*ActionMenuItemView menuItem2 = findViewById(R.id.favoriteIcon);*/
+        ActionMenuItemView menuItem3 = findViewById(R.id.registerIcon);
         if (fragNum == 4) {
             Fragment settingFragment = new SettingFragment();
-            //  fragmentTransaction.replace(R.id.fragment_container, (Fragment)settingFragment).commit();
-
-
             getFragmentManager().beginTransaction().replace(R.id.fragment_container, settingFragment).show(settingFragment).commit();
+            emptyTV.setVisibility(View.GONE);
+
+            menuItem1.setVisibility(View.INVISIBLE);
+           // menuItem2.setVisibility(View.INVISIBLE);
+            menuItem3.setVisibility(View.INVISIBLE);
         } else {
-//            Bundle bundle1 = new Bundle();
+            fragment.setToolbarHandler(toolbarHandler);
+            Bundle bundle = new Bundle();
             bundle.putInt("frag", fragNum);
             fragment.setArguments(bundle);
-            Log.d("nasdfasdfsdfnnn","졸라망함............2");
             fragment.getListDatas();
             getFragmentManager().beginTransaction().replace(R.id.fragment_container, fragment).commit();
+
+            menuItem1.setVisibility(View.VISIBLE);
+          //  menuItem2.setVisibility(View.VISIBLE);
+            menuItem3.setVisibility(View.VISIBLE);
         }
 
 
@@ -237,21 +233,31 @@ public class MainActivity extends android.support.v7.app.AppCompatActivity imple
                         public void onClick(DialogInterface dialogInterface, int i) {
                             Toast.makeText(getApplicationContext(), "삭제합니다.", Toast.LENGTH_SHORT).show();
                             int num;
-                            if (fragNum == 1) {
-                                for (int a = 0; a < numberlist.size(); a++) {
-                                    num = numberlist.get(a);
-                                    Log.d("number::::::::::", num + "...................");
-                                    Log.d("fragment", fragment + "..");
-                                    userCosmeticDAO.deleteItemById(num);
-                                    fragment = MyItemList.newInstance();
-                                    fragment.setToolbarHandler(toolbarHandler);
-                                    Bundle bundle1 = new Bundle();
-                                    bundle1.putInt("frag", fragNum);
-                                    fragment.setArguments(bundle1);
-                                    fragment.getListDatas();
-                                    getFragmentManager().beginTransaction().replace(R.id.fragment_container, fragment).commit();
+                            if (fragNum == 1 || fragNum == 2 || fragNum == 3) {
+                                    for (int a = 0; a < numberlist.size(); a++) {
+                                        num = numberlist.get(a);
+                                        Log.d("number::::::::::", num + "...................");
+                                        Log.d("fragment", fragment + "..");
+                                        if(fragNum == 1) {
+                                            userCosmeticDAO.deleteItemById(num);
+                                        }else  if(fragNum == 2){
+                                            userCosmeticToolsDAO.deleteItemById(num);
+                                        }else if(fragNum == 3){
+                                            userLensDAO.deleteItemById(num);
+                                        }
 
-                                }
+                                        fragment = MyItemList.newInstance();
+                                        fragment.setToolbarHandler(toolbarHandler);
+
+                                        Bundle bundle1 = new Bundle();
+                                        bundle1.putInt("frag", fragNum);
+                                        fragment.setArguments(bundle1);
+                                        Message message = Message.obtain(toolbarHandler,2, null);
+                                        toolbarHandler.sendMessage(message);
+                                        fragment.getListDatas();
+                                        getFragmentManager().beginTransaction().replace(R.id.fragment_container, fragment).commit();
+                                    }
+
 
                             }
                         }
@@ -266,8 +272,8 @@ public class MainActivity extends android.support.v7.app.AppCompatActivity imple
                     alertDialog = builder.create();
                     alertDialog.show();
                 }
-            case R.id.favoriteIcon:
-                getFragmentManager().beginTransaction().remove(fragment).commit();
+            //case R.id.favoriteIcon:
+            /*    getFragmentManager().beginTransaction().remove(fragment).commit();
                 fragment = MyItemList.newInstance();
                 fragment.setToolbarHandler(toolbarHandler);
                 Bundle bundle3 = new Bundle();
@@ -279,17 +285,17 @@ public class MainActivity extends android.support.v7.app.AppCompatActivity imple
                     Log.d("Dddddddddddddd","들어왓음1");
                     getFragmentManager().beginTransaction().replace(R.id.fragment_container, fragment).commit();
                 } else if (favorite == 1) {
-                 /*   fragment = MyItemList.newInstance();
+                 *//*   fragment = MyItemList.newInstance();
                     fragment.setToolbarHandler(toolbarHandler);
                     Bundle bundle4 = new Bundle();
                     bundle4.putInt("frag", fragNum);
-                    fragment.setArguments(bundle4);*/
+                    fragment.setArguments(bundle4);*//*
                     fragment.getListDatas();
                     favorite = 0;
                     getFragmentManager().beginTransaction().replace(R.id.fragment_container, fragment).commit();
                 }
                 //getFragmentManager().beginTransaction().replace(R.id.fragment_container, fragment).commit();
-                Log.d("Dddddddddddddd","들어왓음2");
+                Log.d("Dddddddddddddd","들어왓음2");*/
         }
         return super.onOptionsItemSelected(item);
     }
@@ -356,29 +362,55 @@ public class MainActivity extends android.support.v7.app.AppCompatActivity imple
 
         if (showcheckbox == true) {
             getSupportActionBar().setDisplayHomeAsUpEnabled(false);
-            actionBar.setDisplayShowTitleEnabled(false);
+            getSupportActionBar().setDisplayShowTitleEnabled(false);
             getSupportActionBar().setDisplayShowCustomEnabled(true);
             getSupportActionBar().setCustomView(R.layout.all_delete_button);
 
+            Button button = findViewById(R.id.deleteAllbtn);
+          /*  button.setOnClickListener(new View.OnClickListener() {
+                @Override
+                public void onClick(View v) {
+
+                    Message message = Message.obtain(checkedhandler,1, null);
+                    checkedhandler.sendMessage(message);
+                    button.setText("전체 해제");
+
+                }
+            });
+            if (button.getText().equals("전체 해제")){
+                button.setOnClickListener(new View.OnClickListener() {
+                    @Override
+                    public void onClick(View v) {
+                        Message message = Message.obtain(checkedhandler,2, null);
+                        checkedhandler.sendMessage(message);
+                        button.setText("전체 선텍");
+
+                    }
+                });
+            }
+*/
             menu.removeItem(android.R.id.home);
             menu.removeItem(R.id.searchIcon);
             menu.removeItem(R.id.registerIcon);
-            menu.removeItem(R.id.favoriteIcon);
+          //  menu.removeItem(R.id.favoriteIcon);
             android.view.MenuInflater inflater = getMenuInflater();
             inflater.inflate(R.menu.menu_del, menu);
-        } else {
-            actionBar.setDisplayShowCustomEnabled(true);
-            actionBar.setDisplayShowTitleEnabled(true);
-            actionBar.setDisplayHomeAsUpEnabled(true);
+        } else if(showcheckbox == false) {
 
+            getSupportActionBar().setDisplayShowTitleEnabled(true);
+            getSupportActionBar().setDisplayHomeAsUpEnabled(true);
             getSupportActionBar().setDisplayShowCustomEnabled(false);
-            actionBar.setDisplayShowTitleEnabled(true);
 
+
+            drawerLayout = findViewById(R.id.drawerlayout);
+            ActionBarDrawerToggle toggle = new ActionBarDrawerToggle(this, drawerLayout, toolbar, R.string.openT, R.string.closeT);
+            drawerLayout.addDrawerListener(toggle);
+            toggle.syncState();
 
             menu.removeItem(R.id.delIcon);
-            android.view.MenuInflater inflater = getMenuInflater();
-            inflater.inflate(R.menu.menu_set, menu);
+
         }
+
         return true;
     }
